+++ source/GCPackages/InstalledApplicationLinux/InstalledApplicationLinux.config.ps1
@@ -1,5 +1,5 @@
 configuration InstalledApplicationLinux {
-    Import-DscResource -ModuleName nxtools #-ModuleVersion 0.3.0
+    Import-DscResource -ModuleName nxtools
 
     node InstalledApplicationLinux {
         GC_InstalledApplicationLinux InstalledApplicationLinux {