+++ tests/kitchen/GCPackages/InstalledApplicationLinux.tests.ps1
@@ -2,7 +2,7 @@ BeforeAll {
     $ProgressPreference = 'SilentlyContinue'
     $ErrorActionPreference = 'Continue'
     $ModulePath = (Join-Path -Path '/tmp/verifier' -ChildPath 'modules')
-    $packageZipPath = Join-Path -Path $ModulePath -ChildPath 'GCPolicyPackages/InstalledApplicationLinux*.zip'
+    $packageZipPath = Join-Path -Path $ModulePath -ChildPath 'GCPackages/InstalledApplicationLinux*.zip'
     $packageZip = Get-Item -Path $packageZipPath -errorAction SilentlyContinue
 }
 Describe 'Test InstalledApplicationLinux Audit Package' {
@@ -14,14 +14,14 @@ Describe 'Test InstalledApplicationLinux Audit Package' {
         { Install-GuestConfigurationPackage -Force -Path $packageZip } | Should -not -Throw
     }
 
-    it 'Gets the InstalledApplicationLinux ''powershell-preview'' Package Compliance Status (with params)' {
+    it 'Gets the InstalledApplicationLinux ''powershell'' Package Compliance Status (with params)' {
 
         $result = $null
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip -Parameter @{
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip -Parameter @{
             ResourceType = "GC_InstalledApplicationLinux"
             ResourceId = "InstalledApplicationLinux"
             ResourcePropertyName =  "AttributesYmlContent"
-            ResourcePropertyValue = "packages: [powershell-preview]"
+            ResourcePropertyValue = "packages: [powershell]"
         }
 
         $result.Resources.Reasons | Should -BeNullOrEmpty
@@ -31,11 +31,11 @@ Describe 'Test InstalledApplicationLinux Audit Package' {
     it 'Gets the non-compliant InstalledApplicationLinux @(''powershell-preview'',''somethingNotInstalled'') Package Compliance Status (with params)' {
 
         $result = $null
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip -Parameter @{
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip -Parameter @{
             ResourceType = "GC_InstalledApplicationLinux"
             ResourceId = "InstalledApplicationLinux"
             ResourcePropertyName =  "AttributesYmlContent"
-            ResourcePropertyValue = "powershell-preview;somethingNotInstalled"
+            ResourcePropertyValue = "powershell;somethingNotInstalled"
         }
 
         $result.Resources.Reasons | Should -not -BeNullOrEmpty