+++ tests/kitchen/GCPackages/NotInstalledApplicationLinux.tests.ps1
@@ -2,7 +2,7 @@ BeforeAll {
     $ProgressPreference = 'SilentlyContinue'
     $ErrorActionPreference = 'Continue'
     $ModulePath = (Join-Path -Path '/tmp/verifier' -ChildPath 'modules')
-    $packageZipPath = Join-Path -Path $ModulePath -ChildPath 'GCPolicyPackages/NotInstalledApplicationLinux*.zip'
+    $packageZipPath = Join-Path -Path $ModulePath -ChildPath 'GCPackages/NotInstalledApplicationLinux*.zip'
     $packageZip = Get-Item -Path $packageZipPath -errorAction SilentlyContinue
 }
 Describe 'Test NotInstalledApplicationLinux Audit Package' {
@@ -17,7 +17,7 @@ Describe 'Test NotInstalledApplicationLinux Audit Package' {
     it 'Gets the NotInstalledApplicationLinux ''somethingNotInstalled'' Package Compliance Status (with params)' {
 
         $result = $null
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip -Parameter @{
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip -Parameter @{
             ResourceType = "GC_NotInstalledApplicationLinux"
             ResourceId = "NotInstalledApplicationLinux"
             ResourcePropertyName =  "AttributesYmlContent"
@@ -28,14 +28,14 @@ Describe 'Test NotInstalledApplicationLinux Audit Package' {
         $result.complianceStatus | Should -be $true
     }
 
-    it 'Gets the non-compliant NotInstalledApplicationLinux @(''powershell-preview'',''somethingNotInstalled'') Package Compliance Status (with params)' {
+    it 'Gets the non-compliant NotInstalledApplicationLinux @(''powershell'',''somethingNotInstalled'') Package Compliance Status (with params)' {
 
         $result = $null
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip -Parameter @{
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip -Parameter @{
             ResourceType = "GC_NotInstalledApplicationLinux"
             ResourceId = "NotInstalledApplicationLinux"
             ResourcePropertyName =  "AttributesYmlContent"
-            ResourcePropertyValue = "powershell-preview;somethingNotInstalled"
+            ResourcePropertyValue = "powershell;somethingNotInstalled"
         }
 
         $result.Resources.Reasons | Should -not -BeNullOrEmpty