+++ tests/kitchen/GCPackages/PasswordPolicy_msid232.tests.ps1
@@ -2,7 +2,7 @@ BeforeAll {
     $ProgressPreference = 'SilentlyContinue'
     $ErrorActionPreference = 'Continue'
     $ModulePath = (Join-Path -Path '/tmp/verifier' -ChildPath 'modules')
-    $packageZipPath = Join-Path -Path $ModulePath -ChildPath 'GCPolicyPackages/PasswordPolicy_msid232*.zip'
+    $packageZipPath = Join-Path -Path $ModulePath -ChildPath 'GCPackages/PasswordPolicy_msid232*.zip'
     $packageZip = Get-Item -Path $packageZipPath -errorAction SilentlyContinue
 }
 
@@ -27,7 +27,7 @@ Describe 'Password Policy msid23.2 GC Package (no accounts without passwords)' {
         }
 
         $result = $null
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip
         $result.Resources.Reasons | Should -BeNullOrEmpty
     }
 
@@ -39,7 +39,7 @@ Describe 'Password Policy msid23.2 GC Package (no accounts without passwords)' {
 
         $null = passwd -d test2 *>&1
         $result = $null
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip
         $result.Resources.Reasons | Should -not -BeNullOrEmpty
     }
 }