+++ tests/kitchen/GCPackages/linuxGroupsMustExclude.tests.ps1
@@ -2,10 +2,10 @@ BeforeAll {
     $ProgressPreference = 'SilentlyContinue'
     $ErrorActionPreference = 'Continue'
     $ModulePath = (Join-Path -Path '/tmp/verifier' -ChildPath 'modules')
-    $packageZipPath = Join-Path -Path $ModulePath -ChildPath 'GCPolicyPackages/LinuxGroupsMustExclude*.zip'
+    $packageZipPath = Join-Path -Path $ModulePath -ChildPath 'GCPackages/LinuxGroupsMustExclude*.zip'
     $packageZip = Get-Item -Path $packageZipPath -errorAction SilentlyContinue
 }
-Describe 'Test LinuxGroupsMustExcludt Audit Package' {
+Describe 'Test LinuxGroupsMustExclude Package' {
     it 'Package should be available' {
 
         Test-Path -Path $packageZip | Should -be $true -because (gci (split-path -parent $packageZipPath))
@@ -21,10 +21,10 @@ Describe 'Test LinuxGroupsMustExcludt Audit Package' {
         }
 
         $result = $null
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip
-        $result.Resources.GroupName | Should -be 'foobar'
-        $result.Resources.Ensure | Should -be 'Absent'
-        $result.Resources.Reasons.Phrase | Should -Match "'foobar' was not found but was expected" -Because $result.Reasons
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip
+        $result.Resources.Properties.GroupName | Should -be 'foobar'
+        $result.Resources.Properties.Ensure | Should -be 'Absent'
+        $result.Resources.Properties.Reasons.Phrase | Should -Match "'foobar' was not found but was expected" -Because $result.Reasons
     }
 
     it 'finds the newly created group ''foobar'' to be compliant' {
@@ -33,11 +33,11 @@ Describe 'Test LinuxGroupsMustExcludt Audit Package' {
             New-nxLocalGroup -GroupName foobar -ErrorAction SilentlyContinue -Confirm:$false
         }
 
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip
         $result.complianceStatus | Should -be $true
-        $result.Resources.GroupName | Should -be 'foobar'
-        $result.Resources.Ensure | Should -be 'Present'
-        $result.Resources.Reasons | Should -BeNullOrEmpty
+        $result.Resources.Properties.GroupName | Should -be 'foobar'
+        $result.Resources.Properties.Ensure | Should -be 'Present'
+        $result.Resources.Properties.Reasons | Should -BeNullOrEmpty
     }
 
     it 'finds the newly created group ''foobar'' to be compliant with a parameter' {
@@ -46,7 +46,7 @@ Describe 'Test LinuxGroupsMustExcludt Audit Package' {
             New-nxLocalGroup -GroupName foobar -ErrorAction SilentlyContinue -Confirm:$false
         }
 
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip -Parameter @{
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip -Parameter @{
             ResourceType = "GC_LinuxGroup"
             ResourceId = "LinuxGroupsMustExclude"
             ResourcePropertyName =  "MembersToExcludeAsString"
@@ -54,29 +54,34 @@ Describe 'Test LinuxGroupsMustExcludt Audit Package' {
         }
 
         $result.complianceStatus | Should -be $true
-        $result.Resources.GroupName | Should -be 'foobar'
-        $result.Resources.Ensure | Should -be 'Present'
-        $result.Resources.Reasons | Should -BeNullOrEmpty
+        $result.Resources.Properties.GroupName | Should -be 'foobar'
+        $result.Resources.Properties.Ensure | Should -be 'Present'
+        $result.Resources.Properties.Reasons | Should -BeNullOrEmpty
     }
 
     it 'finds the the group ''foobar'' not compliant when we add root & test users.' {
         if (-not (Get-nxLocalGroup -GroupName 'foobar'))
         {
-            New-nxLocalGroup -GroupName foobar -ErrorAction SilentlyContinue -Confirm:$false
+            New-nxLocalGroup -GroupName 'foobar' -ErrorAction SilentlyContinue -Confirm:$false
+        }
+
+        if (-not (Get-nxLocalUser -UserName 'test'))
+        {
+            New-nxLocalUser -UserName 'test' -Confirm:$false -ErrorAction 'SilentlyContinue'
         }
 
         Add-nxLocalGroupMember -GroupName 'foobar' -UserName 'root','test'
 
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip
         $result.complianceStatus | Should -be $false
-        $result.Resources.GroupName | Should -be 'foobar'
-        $result.Resources.Ensure | Should -be 'Present'
-        $result.Resources.Reasons | Should -not -BeNullOrEmpty
+        $result.Resources.Properties.GroupName | Should -be 'foobar'
+        $result.Resources.Properties.Ensure | Should -be 'Present'
+        $result.Resources.Properties.Reasons | Should -not -BeNullOrEmpty
     }
 
-    it 'Remediates the LinuxGroupsMustInclude package by creating the group ''foobar'' with no params' -skip { # Skipping because package is not yet set as AuditAndSet
-        Start-GuestConfigurationPackageRemediation -Package $packageZip
-        $result = Get-GuestConfigurationPackageComplianceStatus -Package $packageZip -verbose
+    it 'Remediates the LinuxGroupsMustExclude package by creating the group ''foobar'' with no params' { # Skipping because package is not yet set as AuditAndSet
+        Start-GuestConfigurationPackageRemediation -Path $packageZip -Verbose -WarningAction SilentlyContinue
+        $result = Get-GuestConfigurationPackageComplianceStatus -Path $packageZip
         $result.ComplianceStatus | Should -be $true
     }
 }